VB-4332 - Email notifications for public users with gov notify
Initial commit. Introduce new configurations for email and
restructure sms into their own packages for tests etc.
