+++ src/main/resources/application.yml
@@ -81,9 +81,10 @@ hmpps:
 
 notify:
   apikey: ${notify.api.key}
-  sms:
-    templates:
-      VISIT_BOOKING: 85904166-e539-43f5-9f51-7ba106cc61bd
-      VISIT_UPDATE: 386e83ff-5734-4d99-8279-b3eacb7cc8b8
-      VISIT_CANCEL: 42a995f2-abbc-474b-8563-ca2995529111
-      VISIT_CANCEL_NO_PRISON_NUMBER: 3103b319-267d-4265-83a6-a38e93fc2342
+  sms-templates:
+    VISIT_BOOKING: 85904166-e539-43f5-9f51-7ba106cc61bd
+    VISIT_UPDATE: 386e83ff-5734-4d99-8279-b3eacb7cc8b8
+    VISIT_CANCEL: 42a995f2-abbc-474b-8563-ca2995529111
+    VISIT_CANCEL_NO_PRISON_NUMBER: 3103b319-267d-4265-83a6-a38e93fc2342
+  email-templates:
+    VISIT_BOOKING: 85904166-e539-43f5-9f51-7ba106cc61bd # TODO: Replace with real email gov notify template ID.
