+++ src/test/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/integration/domainevents/email/PrisonVisitBookedEventEmailTest.kt
@@ -1,0 +1,5 @@
+package uk.gov.justice.digital.hmpps.notificationsalertsvsip.integration.domainevents.email
+
+// TODO VB-4332: Imeplement tests
+
+class PrisonVisitBookedEventEmailTest
