+++ src/test/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/integration/domainevents/sms/PrisonVisitBookedEventSmsTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.notificationsalertsvsip.integration.domainevents
+package uk.gov.justice.digital.hmpps.notificationsalertsvsip.integration.domainevents.sms
 
 import org.awaitility.kotlin.await
 import org.awaitility.kotlin.untilAsserted
@@ -14,14 +14,15 @@ import uk.gov.justice.digital.hmpps.notificationsalertsvsip.dto.visit.scheduler.
 import uk.gov.justice.digital.hmpps.notificationsalertsvsip.dto.visit.scheduler.VisitDto
 import uk.gov.justice.digital.hmpps.notificationsalertsvsip.enums.SmsTemplateNames
 import uk.gov.justice.digital.hmpps.notificationsalertsvsip.enums.VisitEventType
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.service.listeners.notifiers.PRISON_VISIT_CHANGED
+import uk.gov.justice.digital.hmpps.notificationsalertsvsip.integration.domainevents.EventsIntegrationTestBase
+import uk.gov.justice.digital.hmpps.notificationsalertsvsip.service.listeners.notifiers.PRISON_VISIT_BOOKED
 import java.time.Duration
 import java.time.LocalDate
 import java.time.LocalTime
 import java.time.format.DateTimeFormatter
 import java.time.temporal.ChronoUnit
 
-class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
+class PrisonVisitBookedEventSmsTest : EventsIntegrationTestBase() {
   companion object {
     const val EXPECTED_DATE_PATTERN = "d MMMM yyyy"
   }
@@ -91,15 +92,17 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
   }
 
   @Test
-  fun `when visit updated message is received then update message is sent`() {
+  fun `when visit booked message is received then booking message is sent`() {
     // Given
     val bookingReference = visit.reference
-    val domainEvent = createDomainEventJson(PRISON_VISIT_CHANGED, createAdditionalInformationJson(bookingReference))
+
+    val domainEvent = createDomainEventJson(PRISON_VISIT_BOOKED, createAdditionalInformationJson(bookingReference))
     val jsonSqsMessage = createSQSMessage(domainEvent)
+
+    val templateId = templatesConfig.smsTemplates[SmsTemplateNames.VISIT_BOOKING.name]
     val visitDate = visit.startTimestamp.toLocalDate()
     val expectedVisitDate = visitDate.format(DateTimeFormatter.ofPattern(EXPECTED_DATE_PATTERN))
     val expectedDayOfWeek = visitDate.dayOfWeek.toString().lowercase().replaceFirstChar { it.titlecase() }
-    val templateId = smsTemplatesConfig.templates[SmsTemplateNames.VISIT_UPDATE.name]
     val templateVars = mutableMapOf<String, Any>(
       "prison" to prison.prisonName,
       "time" to "10:30am",
@@ -114,9 +117,9 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
     prisonRegisterMockServer.stubGetPrison(prison.prisonId, prison)
 
     // Then
-    await untilAsserted { verify(prisonVisitChangedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.UPDATED, "bi-vn-wn-ml") }
-    await untilAsserted { verify(smsSenderService, times(1)).sendSms(visit, VisitEventType.UPDATED) }
+    await untilAsserted { verify(prisonVisitBookedEventNotifierSpy, times(1)).processEvent(any()) }
+    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.BOOKED, "bi-vn-wn-ml") }
+    await untilAsserted { verify(smsSenderService, times(1)).sendSms(visit, VisitEventType.BOOKED) }
     await untilAsserted {
       verify(notificationClient, times(1)).sendSms(
         templateId,
@@ -128,15 +131,52 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
   }
 
   @Test
-  fun `when visit updated message is received then update message is sent with th right time format when start time minutes is 00`() {
+  fun `when visit booked message is received then booking message is sent in the right time format when start time minutes is 01`() {
+    // Given
+    val bookingReference = visit3.reference
+    val domainEvent = createDomainEventJson(PRISON_VISIT_BOOKED, createAdditionalInformationJson(bookingReference))
+    val jsonSqsMessage = createSQSMessage(domainEvent)
+
+    // When
+    domainEventListenerService.onDomainEvent(jsonSqsMessage)
+    visitSchedulerMockServer.stubGetVisit(bookingReference, visit3)
+    prisonRegisterMockServer.stubGetPrison(prison.prisonId, prison)
+    val visitDate = visit3.startTimestamp.toLocalDate()
+    val expectedVisitDate = visitDate.format(DateTimeFormatter.ofPattern(EXPECTED_DATE_PATTERN))
+    val templateId = templatesConfig.smsTemplates[SmsTemplateNames.VISIT_BOOKING.name]
+    val expectedDayOfWeek = visitDate.dayOfWeek.toString().lowercase().replaceFirstChar { it.titlecase() }
+    val templateVars = mutableMapOf<String, Any>(
+      "prison" to prison.prisonName,
+      "time" to "12:01am",
+      "dayofweek" to expectedDayOfWeek,
+      "date" to expectedVisitDate,
+      "ref number" to bookingReference,
+    )
+
+    // Then
+    await untilAsserted { verify(prisonVisitBookedEventNotifierSpy, times(1)).processEvent(any()) }
+    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.BOOKED, "qq-yy-xx-kk") }
+    await untilAsserted { verify(smsSenderService, times(1)).sendSms(visit3, VisitEventType.BOOKED) }
+    await untilAsserted {
+      verify(notificationClient, times(1)).sendSms(
+        templateId,
+        visit3.visitContact.telephone,
+        templateVars,
+        visit3.reference,
+      )
+    }
+  }
+
+  @Test
+  fun `when visit booked message is received then booking message is sent in the right time format when start time minutes is 00`() {
     // Given
     val bookingReference = visit2.reference
-    val domainEvent = createDomainEventJson(PRISON_VISIT_CHANGED, createAdditionalInformationJson(bookingReference))
+    val domainEvent = createDomainEventJson(PRISON_VISIT_BOOKED, createAdditionalInformationJson(bookingReference))
     val jsonSqsMessage = createSQSMessage(domainEvent)
     val visitDate = visit2.startTimestamp.toLocalDate()
     val expectedVisitDate = visitDate.format(DateTimeFormatter.ofPattern(EXPECTED_DATE_PATTERN))
     val expectedDayOfWeek = visitDate.dayOfWeek.toString().lowercase().replaceFirstChar { it.titlecase() }
-    val templateId = smsTemplatesConfig.templates[SmsTemplateNames.VISIT_UPDATE.name]
+    val templateId = templatesConfig.smsTemplates[SmsTemplateNames.VISIT_BOOKING.name]
     val templateVars = mutableMapOf<String, Any>(
       "prison" to prison.prisonName,
       "time" to "8am",
@@ -151,9 +191,9 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
     prisonRegisterMockServer.stubGetPrison(prison.prisonId, prison)
 
     // Then
-    await untilAsserted { verify(prisonVisitChangedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.UPDATED, "zz-yy-xx-kk") }
-    await untilAsserted { verify(smsSenderService, times(1)).sendSms(visit2, VisitEventType.UPDATED) }
+    await untilAsserted { verify(prisonVisitBookedEventNotifierSpy, times(1)).processEvent(any()) }
+    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.BOOKED, "zz-yy-xx-kk") }
+    await untilAsserted { verify(smsSenderService, times(1)).sendSms(visit2, VisitEventType.BOOKED) }
     await untilAsserted {
       verify(notificationClient, times(1)).sendSms(
         templateId,
@@ -165,47 +205,10 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
   }
 
   @Test
-  fun `when visit updated message is received then update message is sent with th right time format when start time minutes is 01`() {
-    // Given
-    val bookingReference = visit3.reference
-    val domainEvent = createDomainEventJson(PRISON_VISIT_CHANGED, createAdditionalInformationJson(bookingReference))
-    val jsonSqsMessage = createSQSMessage(domainEvent)
-    val visitDate = visit3.startTimestamp.toLocalDate()
-    val expectedVisitDate = visitDate.format(DateTimeFormatter.ofPattern(EXPECTED_DATE_PATTERN))
-    val expectedDayOfWeek = visitDate.dayOfWeek.toString().lowercase().replaceFirstChar { it.titlecase() }
-    val templateId = smsTemplatesConfig.templates[SmsTemplateNames.VISIT_UPDATE.name]
-    val templateVars = mutableMapOf<String, Any>(
-      "prison" to prison.prisonName,
-      "time" to "12:01am",
-      "dayofweek" to expectedDayOfWeek,
-      "date" to expectedVisitDate,
-      "ref number" to bookingReference,
-    )
-
-    // When
-    domainEventListenerService.onDomainEvent(jsonSqsMessage)
-    visitSchedulerMockServer.stubGetVisit(bookingReference, visit3)
-    prisonRegisterMockServer.stubGetPrison(prison.prisonId, prison)
-
-    // Then
-    await untilAsserted { verify(prisonVisitChangedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.UPDATED, "qq-yy-xx-kk") }
-    await untilAsserted { verify(smsSenderService, times(1)).sendSms(visit3, VisitEventType.UPDATED) }
-    await untilAsserted {
-      verify(notificationClient, times(1)).sendSms(
-        templateId,
-        visit3.visitContact.telephone,
-        templateVars,
-        visit3.reference,
-      )
-    }
-  }
-
-  @Test
-  fun `when visit updated message is received but the visit could not be found then update message is not sent`() {
+  fun `when visit booked message is received but the visit could not be found then booking message is not sent`() {
     // Given
     val bookingReference = visit.reference
-    val domainEvent = createDomainEventJson(PRISON_VISIT_CHANGED, createAdditionalInformationJson(bookingReference))
+    val domainEvent = createDomainEventJson(PRISON_VISIT_BOOKED, createAdditionalInformationJson(bookingReference))
     val jsonSqsMessage = createSQSMessage(domainEvent)
 
     // When
@@ -214,16 +217,16 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
     prisonRegisterMockServer.stubGetPrison(prison.prisonId, prison)
 
     // Then
-    await untilAsserted { verify(prisonVisitChangedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.UPDATED, "bi-vn-wn-ml") }
+    await untilAsserted { verify(prisonVisitBookedEventNotifierSpy, times(1)).processEvent(any()) }
+    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.BOOKED, "bi-vn-wn-ml") }
     await untilAsserted { verify(smsSenderService, times(0)).sendSms(any(), any()) }
   }
 
   @Test
-  fun `when visit updated message is received but the visit is in the past then update message is not sent`() {
+  fun `when visit booked message is received but the visit is in the past then booking message is not sent`() {
     // Given
     val bookingReference = pastDatedVisit.reference
-    val domainEvent = createDomainEventJson(PRISON_VISIT_CHANGED, createAdditionalInformationJson(bookingReference))
+    val domainEvent = createDomainEventJson(PRISON_VISIT_BOOKED, createAdditionalInformationJson(bookingReference))
     val jsonSqsMessage = createSQSMessage(domainEvent)
 
     // When
@@ -232,16 +235,16 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
     prisonRegisterMockServer.stubGetPrison(prison.prisonId, prison)
 
     // Then
-    await untilAsserted { verify(prisonVisitChangedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.UPDATED, "aa-bb-cc-dd") }
+    await untilAsserted { verify(prisonVisitBookedEventNotifierSpy, times(1)).processEvent(any()) }
+    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.BOOKED, "aa-bb-cc-dd") }
     await untilAsserted { verify(smsSenderService, times(0)).sendSms(any(), any()) }
   }
 
   @Test
-  fun `when visit updated message is received but no visit contact then update message is not sent`() {
+  fun `when visit booked message is received but no visit contact found then booking message is not sent`() {
     // Given
     val bookingReference = noContactVisit.reference
-    val domainEvent = createDomainEventJson(PRISON_VISIT_CHANGED, createAdditionalInformationJson(bookingReference))
+    val domainEvent = createDomainEventJson(PRISON_VISIT_BOOKED, createAdditionalInformationJson(bookingReference))
     val jsonSqsMessage = createSQSMessage(domainEvent)
 
     // When
@@ -250,22 +253,23 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
     prisonRegisterMockServer.stubGetPrison(prison.prisonId, prison)
 
     // Then
-    await untilAsserted { verify(prisonVisitChangedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.UPDATED, "bb-cc-dd-zz") }
+    await untilAsserted { verify(prisonVisitBookedEventNotifierSpy, times(1)).processEvent(any()) }
+    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.BOOKED, "bb-cc-dd-zz") }
     await untilAsserted { verify(smsSenderService, times(0)).sendSms(any(), any()) }
   }
 
   @Test
-  fun `when visit date updated to a single digit date then message is sent out with the right visit date format`() {
+  fun `when single digit visit date booked then message is sent out with the right visit date format`() {
     // Given
     val bookingReference = singleDigitDateVisit.reference
-    val domainEvent = createDomainEventJson(PRISON_VISIT_CHANGED, createAdditionalInformationJson(bookingReference))
+    val domainEvent = createDomainEventJson(PRISON_VISIT_BOOKED, createAdditionalInformationJson(bookingReference))
     val jsonSqsMessage = createSQSMessage(domainEvent)
-    val visitYear = singleDigitDateVisit.startTimestamp.toLocalDate().year
+
     // expected visit date should not be 2 digits
+    val visitYear = singleDigitDateVisit.startTimestamp.toLocalDate().year
     val expectedVisitDate = "1 January $visitYear"
     val expectedDayOfWeek = singleDigitDateVisit.startTimestamp.toLocalDate().dayOfWeek.toString().lowercase().replaceFirstChar { it.titlecase() }
-    val templateId = smsTemplatesConfig.templates[SmsTemplateNames.VISIT_UPDATE.name]
+    val templateId = templatesConfig.smsTemplates[SmsTemplateNames.VISIT_BOOKING.name]
     val templateVars = mutableMapOf<String, Any>(
       "prison" to prison.prisonName,
       "time" to "1:05am",
@@ -280,9 +284,9 @@ class PrisonVisitUpdateEventTest : EventsIntegrationTestBase() {
     prisonRegisterMockServer.stubGetPrison(prison.prisonId, prison)
 
     // Then
-    await untilAsserted { verify(prisonVisitChangedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.UPDATED, "bb-cc-dd-xd") }
-    await untilAsserted { verify(smsSenderService, times(1)).sendSms(singleDigitDateVisit, VisitEventType.UPDATED) }
+    await untilAsserted { verify(prisonVisitBookedEventNotifierSpy, times(1)).processEvent(any()) }
+    await untilAsserted { verify(notificationService, times(1)).sendMessage(VisitEventType.BOOKED, "bb-cc-dd-xd") }
+    await untilAsserted { verify(smsSenderService, times(1)).sendSms(singleDigitDateVisit, VisitEventType.BOOKED) }
     await untilAsserted {
       verify(notificationClient, times(1)).sendSms(
         templateId,
